+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### Features
 - [#1875](https://github.com/influxdb/influxdb/pull/1875): Support trace logging of Raft.
+- [#1895](https://github.com/influxdb/influxdb/pull/1895): Auto-create a retention policy when a database is created.
 
 ## v0.9.0-rc9 [2015-03-06]